move methods to World (except take_component)
